Retrieve realm roles (general roles) as well
The roles method currently only retrieve the client roles that a user has been assigned, but not the general roles or realm roles present on the access token payload.
I don't know if there is any good reason to not include the realm roles.

realm_access: Contains a list of global roles. it is the intersection between the roles granted to the user, and the roles the client has access to.
resource_access: This contains a list of client roles.

According the definition of each key on the access token, it will be also better to compare if a user has a role in the real access, I found this as a issue on an application that i'm developing and clearing global roles and not being able to access them in my application.